eCommerce-master/eCommerce/requirements.txt
asgiref==3.2.10
backports.shutil-get-terminal-size==1.0.0
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
cryptography==3.2.1
decorator==4.0.10
Django==3.1.2
django-paypal==1.0.0
django-resized==0.3.11
django-rest-knox==4.1.0
djangorestframework==3.12.2
enum34==1.1.6
idna==2.10
ipython-genutils==0.1.0
pathlib2==2.1.0
paytmchecksum==1.7.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==8.0.1
prompt-toolkit==1.0.9
ptyprocess==0.5.1
pycparser==2.20
pycryptodome==3.9.9
PyMySQL==0.10.1
pytz==2020.1
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
wcwidth==0.1.7
